WIP: log: convert log functions to macros
*Changes:*

* Always assume the log function should not be called this improves
production performance at the cost of debug performance.  In production
we only log WARN/ERROR so performance doesn't matter when we have to log
since the caller is already in a bad state.

* Mark all log functions as "never inline"

* Add the format attribute to log functions to improve compiler
warnings.

NOTE: This takes advantage of the GCC variadic macro extension, but
works with Clang (though a warning is emitted).