Handle literal use of "status" on statsd server
Previous versions would report their metrics when interrogating with
the word status, which is pretty silly, but the current version will
try to forward an un-typed "status" frame upstream.

This special cases the word status and throws away any content
containing it, so as not to send status to any sinks or
processors. Metrics are available on the admin HTTP server.