Support directed propagation from servers to backends
Refactors the servers to support directing at backends, and introduces
direction for processors. All config items now need a route: key which
lists the backend to send it to (e.g., statsd:backend1), or a
processor to send to (e.g. processor:clean_tags). Processors are
unimplemented in this PR.